split the splitting of the HDF5 loader into its own PR to avoid merge hell
